+++ lib/ember-addon/blueprints/ember-data/index.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = {
+  normalizeEntityName: function() {
+    // this prevents an error when the entityName is
+    // not specified (since that doesn't actually matter
+    // to us
+  },
+
+  afterInstall: function() {
+    return this.addBowerPackageToProject('ember-data', '1.0.0-beta.9');
+  }
+};
