+++ lib/ember-addon/index.js
@@ -1,0 +1,26 @@
+'use strict';
+
+var path = require('path');
+
+module.exports = {
+  name: 'ember-data',
+  blueprintsPath: function() {
+    return path.join(__dirname, 'blueprints');
+  },
+  included: function(app) {
+    this._super.included(app);
+
+    var options = {
+      exports: {
+        'ember-data': [
+          'default'
+        ]
+      }
+    };
+
+    this.app.import({
+      development: app.bowerDirectory + '/ember-data/ember-data.js',
+      production: app.bowerDirectory + '/ember-data/ember-data.prod.js'
+    }, options);
+  }
+};
