+++ package.json
@@ -8,13 +8,15 @@
     "ember-addon"
   ],
   "ember-addon": {
-    "main": "lib/ember-addon-main.js"
+    "main": "lib/ember-addon/index.js"
   },
+  "files": [
+    "lib/ember-addon"
+  ],
   "scripts": {
-    "postinstall": "bower install",
-    "prepublish": "ember build --output-path=dist/ember-data/ --environment production",
+    "prepublish": "bower install",
     "start": "ember serve",
-    "dist": "ember build --environment production",
+    "dist": "ember build --environment=production",
     "test": "testem -R dot ci",
     "publish-build": "npm run dist && ./bin/publish_to_s3.js",
     "testem-local": "testem -R dot ci",
